database_service/requirements.txt
dnspython==2.2.1
grpcio==1.45.0
grpcio-tools==1.45.0
protobuf==3.19.4
pymongo==4.1.1
python-dotenv==0.20.0
six==1.16.0
